+++ src/compas_timber/_fabrication/__init__.py
@@ -1,0 +1,11 @@
+# This module is a tepmorary solution to the problem of circular imports in the compas_timber package.
+# It will be reoved or merged with the `fabrication` module once the migration to the new feature system is complete.
+from .btlx_process import BTLxProcess
+from .btlx_process import OrientationType
+from .jack_cut import JackRafterCut
+from .jack_cut import JackRafterCutParams
+from .drilling import Drilling
+from .drilling import DrillingParams
+
+
+__all__ = ["JackRafterCut", "BTLxProcess", "OrientationType", "JackRafterCutParams", "Drilling", "DrillingParams"]